clean up share VariantSearchFormContainer
