+++ ui/pages/Search/components/VariantSearchForm.jsx
@@ -1,46 +1,91 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
+import { Segment } from 'semantic-ui-react'
+import createDecorator from 'final-form-calculate'
 
-import { getLocusListIsLoading } from 'redux/selectors'
-import VariantSearchFormContainer from 'shared/components/panel/search/VariantSearchFormContainer'
+import { navigateSavedHashedSearch } from 'redux/rootReducer'
+import { getSearchedVariantsErrorMessage, getSearchedVariantsIsLoading, getLocusListIsLoading } from 'redux/selectors'
+import FormWrapper from 'shared/components/form/FormWrapper'
+import { toUniqueCsvString } from 'shared/utils/stringUtils'
+import { LOCUS_LIST_ITEMS_FIELD } from 'shared/utils/constants'
+
+import SearchDisplayForm from 'shared/components/panel/search/SearchDisplayForm'
+import { LOCUS_FIELD_NAME, PANEL_APP_FIELD_NAME } from 'shared/components/panel/search/constants'
+
 import { SaveSearchButton } from './SavedSearch'
 import VariantSearchFormContent from './VariantSearchFormContent'
 import { getIntitialSearch, getMultiProjectFamilies } from '../selectors'
 
-const VariantSearchForm = React.memo(({ history, initialSearch, contentLoading, noEditProjects, match }) => (
+const DECORATORS = [
+  createDecorator({
+    field: `search.${LOCUS_FIELD_NAME}.${PANEL_APP_FIELD_NAME}`,
+    updates: {
+      [`search.${LOCUS_FIELD_NAME}.${LOCUS_LIST_ITEMS_FIELD.name}`]: locusValue => (
+        locusValue && toUniqueCsvString(Object.values(locusValue))
+      ),
+    },
+  }),
+]
+
+const VariantSearchForm = React.memo((
+  { initialSearch, contentLoading, noEditProjects, match, onSubmit, resultsPath, variantsLoading, submissionError },
+) => (
   <div>
-    <VariantSearchFormContainer
-      history={history}
+    <FormWrapper
+      onSubmit={onSubmit}
       initialValues={initialSearch}
-      loading={contentLoading}
-      match={match}
+      loading={contentLoading || variantsLoading}
+      submissionError={submissionError}
+      submitButtonText="Search"
+      noModal
+      decorators={DECORATORS}
     >
       <VariantSearchFormContent noEditProjects={noEditProjects} />
       <SaveSearchButton />
-    </VariantSearchFormContainer>
+    </FormWrapper>
+    {!match.params.searchHash && (
+      <Segment basic floated="right"><SearchDisplayForm match={match} /></Segment>
+    )}
   </div>
 ))
 
 VariantSearchForm.propTypes = {
-  history: PropTypes.object,
   match: PropTypes.object,
   initialSearch: PropTypes.object,
   contentLoading: PropTypes.bool,
+  variantsLoading: PropTypes.bool,
   noEditProjects: PropTypes.bool,
+  onSubmit: PropTypes.func,
+  resultsPath: PropTypes.string,
+  submissionError: PropTypes.string,
 }
 
-const mapStateToProps = (state, ownProps) => ({
-  initialSearch: getIntitialSearch(state, ownProps),
+const mapDispatchToProps = (dispatch, ownProps) => ({
+  onSubmit: values => dispatch(navigateSavedHashedSearch(
+    values, ownProps.history.push, ownProps.resultsPath,
+  )),
+})
+
+const getSharedStateToProps = state => ({
+  variantsLoading: getSearchedVariantsIsLoading(state),
+  submissionError: getSearchedVariantsErrorMessage(state),
   contentLoading: getLocusListIsLoading(state),
 })
 
-export default connect(mapStateToProps)(VariantSearchForm)
+const mapStateToProps = (state, ownProps) => ({
+  initialSearch: getIntitialSearch(state, ownProps),
+  ...getSharedStateToProps(state),
+})
+
+export default connect(mapStateToProps, mapDispatchToProps)(VariantSearchForm)
 
 const mapNoEditProjectStateToProps = (state, ownProps) => ({
   noEditProjects: true,
   initialSearch: getMultiProjectFamilies(state, ownProps),
-  contentLoading: getLocusListIsLoading(state),
+  ...getSharedStateToProps(state),
 })
 
-export const NoEditProjectsVariantSearchForm = connect(mapNoEditProjectStateToProps)(VariantSearchForm)
+export const NoEditProjectsVariantSearchForm = connect(
+  mapNoEditProjectStateToProps, mapDispatchToProps,
+)(VariantSearchForm)
